backend/pyproject.toml
[project]
name = "premium-support-highlights"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "deepagents",
    "langchain-anthropic",
    "httpx",
    "python-dotenv",
    "pydantic",
    "beautifulsoup4",
    "langsmith>=0.2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]
